Backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
motor==3.6.0
python-dotenv==1.0.1
bcrypt==4.2.1
pydantic[email]==2.10.3
python-multipart==0.0.20
